For #17303: Move ext fun runIfFragmentIsAttached to utils
move to ext
